% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mash_predict.R
\name{coef.mr.mash}
\alias{coef.mr.mash}
\title{Extract coefficients from mr.mash fit.}
\usage{
\method{coef}{mr.mash}(object)
}
\arguments{
\item{object}{a mr.mash fit.}
}
\value{
(p+1) x r matrix of coefficients.
}
\description{
Extract coefficients from mr.mash fit.
}
