% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mash_rss_predict.R
\name{coef.mr.mash.rss}
\alias{coef.mr.mash.rss}
\title{Extract coefficients from mr.mash.rss fit.}
\usage{
\method{coef}{mr.mash.rss}(object, ...)
}
\arguments{
\item{object}{a mr.mash fit.}

\item{\dots}{Other arguments (not used).}
}
\value{
p x r or (p+1) x r matrix of coefficients,
 depending on whether an intercept was computed.
}
\description{
Extract coefficients from mr.mash.rss fit.
}
