% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{compute_canonical_covs}
\alias{compute_canonical_covs}
\title{Compute canonical covariance matrices.}
\usage{
compute_canonical_covs(
  r,
  singletons = TRUE,
  hetgrid = c(0, 0.25, 0.5, 0.75, 1)
)
}
\arguments{
\item{r}{number of responses.}

\item{singletons}{if \code{TRUE}, the response-specific effect matrices will be 
included.}

\item{hetgrid}{scalar or numeric vector of positive correlation [0, 1] of the effects
across responses.}
}
\value{
A list containing the canonical covariance matrices.
}
\description{
Function to compute canonical covariance matrices scaled by a grid.
}
