% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{compute_cov_canonical}
\alias{compute_cov_canonical}
\title{Compute canonical covariance matrices for \code{mr.mash}.}
\usage{
compute_cov_canonical(
  r,
  singletons = TRUE,
  hetgrid = c(0, 0.25, 0.5, 0.75, 1),
  grid,
  zeromat = TRUE
)
}
\arguments{
\item{r}{number of responses.}

\item{singletons}{if \code{TRUE}, the response-specific effect matrices will be 
included.}

\item{hetgrid}{scalar or numeric vector of positive correlation [0, 1] of the effects
across responses.}

\item{grid}{scalar or numeric vector of variances of the effects.}

\item{zeromat}{if \code{TRUE}, the no-effect matrix will be included.}
}
\value{
A list containing the canonical covariance matrices.
}
\description{
Function to compute canonical covariance matrices scaled by a grid.
}
