% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\name{expand_covs}
\alias{expand_covs}
\title{Expand covariance matrices by a grid of variances for use in \code{mr.mash}.}
\usage{
expand_covs(mats, grid, zeromat = TRUE)
}
\arguments{
\item{mats}{a list of covariance matrices.}

\item{grid}{scalar or numeric vector of variances of the effects.}

\item{zeromat}{if \code{TRUE}, the no-effect matrix will be included.}
}
\value{
A list containing the scaled covariance matrices.
}
\description{
Function to scale the covariance matrices by a grid of variances.
}
