% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mash.R
\name{mr.mash}
\alias{mr.mash}
\title{Multiple Regression with Multivariate Adaptive Shrinkage.}
\usage{
mr.mash(Y, X, V, S0, w0, mu_init = matrix(0, nrow = ncol(X), ncol =
  ncol(Y)), tol = 1e-08, max_iter = 1e+05, update_w0 = T,
  compute_ELBO = T)
}
\arguments{
\item{Y}{an NxR matrix of responses.}

\item{X}{an NxP matrix of covariates.}

\item{V}{an RxR residual covariance matrix.}

\item{S0}{a list of length K containing the desired RxR prior covariance matrices 
on the regression coefficients.}

\item{w0}{a K-vector with prior mixture weights, each associated with the 
respective covariance matrix in \code{S0}.}

\item{mu_init}{a PxR matrix of initial estimates for the regression coefficients.}

\item{tol}{convergence tolerance.}

\item{max_iter}{maximum number of iterations for the optimization algorithm.}

\item{update_w0}{if TRUE, prior weights are updated.}

\item{compute_ELBO}{if TRUE, ELBO is computed.}
}
\value{
a mr.mash fit, which is a list with some or all of the following elements\cr
\item{mu1}{a PxR matrix of posterior means for the regression coeffcients}
\item{S1}{a RxRxP array of posterior covariances for the regression coeffcients}
\item{mu1}{a PxK matrix of posterior assignment probabilities to the mixture components}
\item{intercept}{an R-vector with the estimated intercepts}
\item{ELBO}{the Evidence Lower Bound at convergence}
}
\description{
Multiple Regression with Multivariate Adaptive Shrinkage.
}
\details{
Performs multivariate multiple regression with mixture-of-normals prior.
}
