% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mash_standardized_X.R
\name{mr.mash.scaled.X}
\alias{mr.mash.scaled.X}
\title{Multiple Regression with Multivariate Adaptive Shrinkage with scaled X.}
\usage{
mr.mash.scaled.X(
  Y,
  X,
  V,
  S0,
  w0,
  mu_init = matrix(0, nrow = ncol(X), ncol = ncol(Y)),
  tol = 1e-08,
  max_iter = 1e+05,
  update_w0 = T,
  compute_ELBO = T
)
}
\arguments{
\item{Y}{an NxR matrix of responses.}

\item{X}{an NxP matrix of covariates.}

\item{V}{an RxR residual covariance matrix.}

\item{S0}{a list of length K containing the desired RxR prior covariance matrices 
on the regression coefficients.}

\item{w0}{a K-vector with prior mixture weights, each associated with the 
respective covariance matrix in \code{S0}.}

\item{mu_init}{a PxR matrix of initial estimates for the regression coefficients.}

\item{tol}{convergence tolerance.}

\item{max_iter}{maximum number of iterations for the optimization algorithm.}

\item{update_w0}{if TRUE, prior weights are updated.}

\item{compute_ELBO}{if TRUE, ELBO is computed.}
}
\value{
a mr.mash fit, which is a list with some or all of the following elements\cr
\item{mu1}{a PxR matrix of posterior means for the regression coeffcients}
\item{S1}{a RxRxP array of posterior covariances for the regression coeffcients}
\item{w1}{a PxK matrix of posterior assignment probabilities to the mixture components}
\item{intercept}{an R-vector with the estimated intercepts}
\item{ELBO}{the Evidence Lower Bound at convergence}
}
\description{
Multiple Regression with Multivariate Adaptive Shrinkage with scaled X.
}
\details{
Performs multivariate multiple regression with mixture-of-normals prior.
}
\examples{
###Set options
options(stringsAsFactors = F)
###Set seed
set.seed(123)

###Simulate X and Y
##Set parameters
n  <- 100
p <- 10

##Compute residual covariance
V  <- rbind(c(1.0,0.2),
            c(0.2,0.4))

##Set true effects
B  <- matrix(c(-2, -2,
               5, 5,
               rep(0, (p-2)*2)), byrow=T, ncol=2)

##Simulate X
X <- matrix(rnorm(n*p), nrow=n, ncol=p)
X <- scale(X, center=T, scale=F)

##Simulate Y from MN(XB, I_n, V) where I_n is an nxn identity matrix and V is the residual covariance  
Y <- mr.mash.alpha:::sim_mvr(X, B, V)

###Specify the mixture weights and covariance matrices for the mixture-of-normals prior.
grid <- seq(1, 5)
S0mix <- mr.mash.alpha:::compute_cov_canonical(ncol(Y), singletons=T, hetgrid=c(0, 0.25, 0.5, 0.75, 0.99), grid, zeromat=T)
w0    <- rep(1/(length(S0mix)), length(S0mix))

###Estimate residual covariance
V_est <- cov(Y)

###Fit mr.mash
fit <- mr.mash.scaled.X(Y, X, V_est, S0mix, w0, tol=1e-8, update_w0=T, compute_ELBO=T)

}
