% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_mash_rss_predict.R
\name{predict.mr.mash.rss}
\alias{predict.mr.mash.rss}
\title{Predict future observations from mr.mash.rss fit.}
\usage{
\method{predict}{mr.mash.rss}(object, newx, ...)
}
\arguments{
\item{object}{a mr.mash.rss fit.}

\item{newx}{a new value for X at which to do predictions.}

\item{\dots}{Additional arguments (not used).}
}
\value{
Matrix of predicted values.
}
\description{
Predict future observations from mr.mash.rss fit.
}
