% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covs.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{estimate_null_correlation_simple}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mashr}{\code{\link[mashr]{estimate_null_correlation_simple}}}
}}

