% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{autoselect.mixsd}
\alias{autoselect.mixsd}
\title{Compute a grid of standard deviations to scale the canonical covariance matrices.}
\usage{
autoselect.mixsd(bhat, shat, mult = 2)
}
\arguments{
\item{bhat}{a numeric vector of regression coefficients.}

\item{shat}{a numeric vector of of standard erros for the regression coefficients.}

\item{mult}{a scalar affecting how dense the resulting grid of standard deviations will be.}
}
\value{
A numeric vector of standard deviations.
}
\description{
Function to compute a grid of standard deviations from univariate 
  simple linear regression summary statistics
}
