% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mr_mash.R
\name{predict.mr.mash}
\alias{predict.mr.mash}
\title{Predict future observations from mr.mash fit.}
\usage{
\method{predict}{mr.mash}(object, newx)
}
\arguments{
\item{object}{a mr.mash fit.}

\item{newx}{a new value for X at which to do predictions.}
}
\value{
a matrix of predicted values.
}
\description{
Predict future observations from mr.mash fit.
}
